+++ src/steps/application/index.ts
@@ -1,0 +1,47 @@
+import {
+  createDirectRelationship,
+  Entity,
+  IntegrationStep,
+  IntegrationStepExecutionContext,
+  RelationshipClass,
+} from '@jupiterone/integration-sdk-core';
+
+import { createAPIClient } from '../../client';
+import { IntegrationConfig } from '../../config';
+import { ACCOUNT_ENTITY_KEY } from '../account';
+import { Entities, Steps, Relationships } from '../constants';
+import { createApplicationEntity } from './converter';
+
+export async function fetchApplications({
+  instance,
+  jobState,
+}: IntegrationStepExecutionContext<IntegrationConfig>) {
+  const apiClient = createAPIClient(instance.config);
+
+  const accountEntity = (await jobState.getData(ACCOUNT_ENTITY_KEY)) as Entity;
+
+  await apiClient.iterateApplications(async (application) => {
+    const applicationEntity = await jobState.addEntity(
+      createApplicationEntity(application),
+    );
+
+    await jobState.addRelationship(
+      createDirectRelationship({
+        _class: RelationshipClass.HAS,
+        from: accountEntity,
+        to: applicationEntity,
+      }),
+    );
+  });
+}
+
+export const applicationSteps: IntegrationStep<IntegrationConfig>[] = [
+  {
+    id: Steps.APPLICATIONS,
+    name: 'Fetch Applications',
+    entities: [Entities.APPLICATION],
+    relationships: [Relationships.ACCOUNT_HAS_APPLICATION],
+    dependsOn: [Steps.ACCOUNT],
+    executionHandler: fetchApplications,
+  },
+];